BIG UPDATE: Black Market
You can now be approached by a black market worker who will vandalize cars for you! Be careful though there is a chance they can rat you out as well!

Made a new class and 2 new methods to execute this idea. Then this is only runs if the playerCount is 1.